requirements.txt
transformers>=4.35.0
torch>=2.0.0
deepspeed>=0.12.0
peft>=0.7.0
Pillow>=9.0.0
torchvision>=0.15.0
wandb>=0.15.0
accelerate>=0.25.0
bitsandbytes>=0.41.3
flash-attn>=2.5.0
importlib-metadata>=6.0.0
scipy>=1.10.0
